batch

+++ iLgs/Models/PsCardItemTransfer.cs
@@ -25,13 +25,13 @@ namespace iLgs.Models
         public System.Guid Id { get; set; }
         public Nullable<System.Guid> PsCardItemId { get; set; }
         public Nullable<System.Guid> ParentId { get; set; }
-        public Nullable<int> Qty { get; set; }
+        public Nullable<decimal> Qty { get; set; }
         public Nullable<System.DateTime> TransDate { get; set; }
         public Nullable<System.Guid> LocationId { get; set; }
-        public Nullable<int> QtyIss { get; set; }
-        public Nullable<int> QtyBal { get; set; }
-        public Nullable<int> TransferIn { get; set; }
-        public Nullable<int> TransferOut { get; set; }
+        public Nullable<decimal> QtyIss { get; set; }
+        public Nullable<decimal> QtyBal { get; set; }
+        public Nullable<decimal> TransferIn { get; set; }
+        public Nullable<decimal> TransferOut { get; set; }
         public string TranType { get; set; }
         public Nullable<decimal> Amount { get; set; }
         public string InsertedBy { get; set; }
