+++ iLgs/Models/RisItem.cs
@@ -31,8 +31,8 @@ namespace iLgs.Models
         public string Description { get; set; }
         public string OtherDesc { get; set; }
         public string Unit { get; set; }
-        public Nullable<int> QtyRequest { get; set; }
-        public Nullable<int> QtyIssue { get; set; }
+        public Nullable<decimal> QtyRequest { get; set; }
+        public Nullable<decimal> QtyIssue { get; set; }
         public string Remarks { get; set; }
         public string InsertedBy { get; set; }
         public Nullable<System.DateTime> InsertedDt { get; set; }
